+++ src/main/java/duke/exceptions/NoThisNumOfTaskException.java
@@ -1,4 +1,4 @@
-package Duke.Exceptions;
+package duke.exceptions;
 
 public class NoThisNumOfTaskException extends DukeException{
     public NoThisNumOfTaskException() {
