+++ src/main/java/duke/tasks/Deadline.java
@@ -1,22 +1,22 @@
-package Duke.Tasks;
+package duke.tasks;
 
-import Duke.Tool.Storage;
-import Duke.Tool.TaskList;
-import Duke.Ui;
+import duke.tool.Storage;
+import duke.tool.TaskList;
+
+import duke.Ui;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class Event extends Task{
-    protected LocalDateTime at;
-
-    public Event(String description, LocalDateTime at, Boolean isDone) {
+public class Deadline extends Task {
+    protected LocalDateTime by;
+    public Deadline(String description, LocalDateTime by, Boolean isDone) {
         super(description, isDone);
-        this.at = at;
+        this.by = by;
     }
 
     @Override
-    public Event markDone() {
+    public Deadline markDone() {
         super.markDone();
         return this;
     }
@@ -28,17 +28,18 @@ public class Event extends Task{
         storage.writeData(tasklist.taskList);
     }
 
+
     @Override
     public String toString() {
-        return "[E]" + super.toString() + " (at: " +
-                at.format(DateTimeFormatter.ofPattern("MMM dd yyyy HH:mm"))  + ")";
+        return "[D]" + super.toString() + " (by: "
+                + by.format(DateTimeFormatter.ofPattern("MMM dd yyyy HH:mm")) + ")";
     }
 
     @Override
     public String fileFormattedString() {
         String doneOrNot = isDone ? "1" : "0";
-        return "E | " + doneOrNot + " | " + this.name + " | " +
-                at.format(DateTimeFormatter.ofPattern("MMM dd yyyy HH:mm"));
+        return "D | " + doneOrNot + " | " + this.name + " | " +
+                by.format(DateTimeFormatter.ofPattern("MMM dd yyyy HH:mm"));
     }
 
 }
