+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
     _______________________________________
-      Hello! I'm Duke yy
+      Hello! I'm duke yy
       What can I do for you?
     _______________________________________
     _______________________________________
